requirements.txt
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.6.2.post1
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
blobfile==3.0.0
certifi==2024.8.30
charset-normalizer==3.4.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
executing==2.1.0
fastjsonschema==2.21.0
filelock==3.16.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==2.1.3
openai==1.55.2
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.3.6
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
pycryptodomex==3.21.0
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sseclient-py==1.7.2
stack-data==0.6.3
tabulate==0.9.0
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
